feat:修改view实现
