+++ lib/route/MainRoute.dart
@@ -3,8 +3,7 @@ import 'package:github/route/main/MainPage.dart';
 import 'package:github/view/bind/BindViewPager.dart';
 import 'package:github/view/main/MainContentsWidget.dart';
 import 'package:github/view/widget/BaseWidget.dart';
-import 'package:github/view/widget/TextView.dart';
-import 'package:github/view/widget/View.dart';
+import 'package:github/view/widget/view/View.dart';
 import 'package:rxdart/rxdart.dart';
 
 class MainRoute extends StatefulWidget {
