+++ lib/route/main/MyPage.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:github/view/widget/TextView.dart';
+import 'package:github/view/widget/view/TextView.dart';
 
 /// 首页homepage
 class MyPage extends StatefulWidget {
