+++ lib/util/DialogManager.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:github/view/widget/BaseWidget.dart';
-import 'package:github/view/widget/View.dart';
+import 'package:github/view/widget/view/View.dart';
 
 class DialogManager {
   static DialogManager _instance;
