+++ lib/util/NaigatorUtils.dart
@@ -23,7 +23,7 @@ class NavigatorUtils {
 
   static Map<String, WidgetBuilder> routeMap = {
     "/": (BuildContext context) => SplashRoute(),
-    "/login": (BuildContext context) => ViewRoute(),
+    "/login": (BuildContext context) => LoginRoute(),
     "/main": (BuildContext context) => ViewRoute(),
   };
