+++ lib/view/main/WBTitleWidget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:github/view/widget/BaseWidget.dart';
-import 'package:github/view/widget/TextView.dart';
-import 'package:github/view/widget/View.dart';
+import 'package:github/view/widget/view/TextView.dart';
+import 'package:github/view/widget/view/View.dart';
 
 class WBTitle extends StatefulWidget {
   final WBTitleOption option;
