+++ lib/view/widget/CustomDialog.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:github/util/util.dart';
 import 'package:github/view/widget/BaseWidget.dart';
+import 'package:github/view/widget/view/View.dart';
 
-import 'View.dart';
 
 class CustomDialog extends StatefulWidget {
   final Widget childWidget;
