+++ lib/view/widget/WebView.dart
@@ -11,7 +11,7 @@ import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:rxdart/rxdart.dart';
 
-import 'TextView.dart';
+import 'view/TextView.dart';
 
 ///
 /// * author : kyle
