+++ lib/view/widget/view/TextView.dart
@@ -2,9 +2,9 @@ import 'package:github/util/util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/framework.dart';
-import 'package:github/view/widget/View.dart';
 
-import 'BaseWidget.dart';
+import 'View.dart';
+
 
 ///
 /// * Author    : kyle
